refactor: replace mobile card layout with ColumnGroupPicker and MobileCompareRows
